Rebased arm-gic fork.
Now you have to tell it which CPU you are on.

+++ cortex-r/src/register/cbar.rs
@@ -21,8 +21,8 @@ impl Cbar {
     }
 
     /// Get the periphbase address
-    pub fn periphbase(self) -> *mut u64 {
-        (self.0 & 0xFFF00000) as *mut u64
+    pub fn periphbase(self) -> *mut u32 {
+        (self.0 & 0xFFF00000) as *mut u32
     }
 }
